refactor: Messenger의 파일명을 변경하고 MESSAGE의 위치를 변경한다
