+++ src/js/constants/index.js
@@ -1,5 +1,4 @@
 import CLASSNAME from "./classname.js";
-import MESSAGE from "./message.js";
 import SNACKBAR_TEXT from "./snackbarText.js";
 import LOCAL_STORAGE_KEY from "./localStorageKey.js";
 import REGULAR_EXPRESSION from "./regularExpression.js";
@@ -7,7 +6,6 @@ import NUMBER from "./number.js";
 
 export {
   CLASSNAME,
-  MESSAGE,
   SNACKBAR_TEXT,
   LOCAL_STORAGE_KEY,
   REGULAR_EXPRESSION,
