+++ src/js/messenger/index.js
@@ -1,0 +1,4 @@
+import MESSAGE from "./message.js";
+import messenger from "./messenger.js";
+
+export { MESSAGE, messenger };
